+++ app/(main)/content_management/layout.tsx
