+++ app/(main)/protected/page.tsx
